+++ cookbooks/nginx/recipes/passenger.rb
@@ -15,9 +15,19 @@
 # limitations under the License.
 #
 
-node.default["nginx"]["passenger"]["version"] = "3.0.12"
-node.default["nginx"]["passenger"]["root"] = "/usr/lib/ruby/gems/1.8/gems/passenger-3.0.12"
-node.default["nginx"]["passenger"]["ruby"] = %x{which ruby}.chomp
+node.default["nginx"]["passenger"]["version"] = "3.0.19"
+
+if node["languages"].attribute?("ruby")
+  node.default["nginx"]["passenger"]["root"] = "#{node['languages']['ruby']['gems_dir']}/gems/passenger-#{node['nginx']['passenger']['version']}"
+  node.default["nginx"]["passenger"]["ruby"] = node['languages']['ruby']['ruby_bin']
+else
+  Chef::Log.warn("node['languages']['ruby'] attribute not detected in #{cookbook_name}::#{recipe_name}")
+  Chef::Log.warn("Install a Ruby for automatic detection of node['nginx']['passenger'] attributes (root, ruby)")
+  Chef::Log.warn("Using default values that may or may not work for this system.")
+  node.default["nginx"]["passenger"]["root"] = "/usr/lib/ruby/gems/1.8/gems/passenger-#{node['nginx']['passenger']['version']}"
+  node.default["nginx"]["passenger"]["ruby"] = "/usr/bin/ruby"
+end
+
 node.default["nginx"]["passenger"]["max_pool_size"] = 10
 node.default["nginx"]["passenger"]["spawn_method"] = "smart-lv2"
 node.default["nginx"]["passenger"]["use_global_queue"] = "on"
@@ -32,7 +42,7 @@ node.default["nginx"]["passenger"]["gem_binary"] = nil
 packages = value_for_platform( ["redhat", "centos", "scientific", "amazon", "oracle"] => {
                                  "default" => %w(ruby-devel curl-devel) },
                                ["ubuntu", "debian"] => {
-                                 "default" => %w(ruby-dev curl-dev) } )
+                                 "default" => %w(ruby-dev libcurl4-gnutls-dev) } )
 
 packages.each do |devpkg|
   package devpkg