+++ cookbooks/runit/libraries/provider_runit_service.rb
@@ -76,7 +76,26 @@ class Chef
         #
         # Chef::Provider::Service overrides
         #
-        def enable_service
+
+        def action_enable
+          converge_by("configure service #{@new_resource}") do
+            configure_service # Do this every run, even if service is already enabled and running
+            Chef::Log.info("#{@new_resource} configured")
+          end
+          if @current_resource.enabled
+            Chef::Log.debug("#{@new_resource} already enabled - nothing to do")
+          else
+            converge_by("enable service #{@new_resource}") do
+              enable_service
+              Chef::Log.info("#{@new_resource} enabled")
+            end
+          end
+          load_new_resource_state
+          @new_resource.enabled(true)
+          restart_service if @new_resource.restart_on_update and run_script.updated_by_last_action?
+        end
+
+        def configure_service
           if new_resource.sv_templates
             Chef::Log.debug("Creating sv_dir for #{new_resource.service_name}")
             sv_dir.run_action(:create)
@@ -120,7 +139,9 @@ class Chef
 
           Chef::Log.debug("Creating lsb_init compatible interface #{new_resource.service_name}")
           lsb_init.run_action(:create)
+        end
 
+        def enable_service
           unless node['platform'] == 'gentoo'
             Chef::Log.debug("Creating symlink in service_dir for #{new_resource.service_name}")
             service_link.run_action(:create)
@@ -161,7 +182,7 @@ class Chef
         #
 
         # only take action if the service is running
-        [:down, :hup, :int, :term, :kill].each do |signal|
+        [:down, :hup, :int, :term, :kill, :quit].each do |signal|
           define_method "action_#{signal}".to_sym do
             if @current_resource.running
               runit_send_signal(signal)
@@ -247,7 +268,7 @@ EOF
           @run_script = Chef::Resource::Template.new(::File.join(sv_dir_name, 'run'), run_context)
           @run_script.owner(new_resource.owner)
           @run_script.group(new_resource.group)
-          @run_script.source("sv-#{new_resource.service_name}-run.erb")
+          @run_script.source("sv-#{new_resource.run_template_name}-run.erb")
           @run_script.cookbook(template_cookbook)
           @run_script.mode(00755)
           if new_resource.options.respond_to?(:has_key?)