+++ cookbooks/wt_actioncenter_dd_responsys/attributes/default.rb
@@ -1,0 +1,14 @@
+#
+# Cookbook Name:: wt_actioncenter_dd_responsys
+# Attributes:: default
+#
+# Copyright 2012, Webtrends
+#
+# All rights reserved - Do Not Redistribute
+#
+
+default['wt_actioncenter_dd_responsys']['user']           = "webtrends"
+default['wt_actioncenter_dd_responsys']['group']          = "webtrends"
+default['wt_actioncenter_dd_responsys']['download_url']   =
+"http://teamcity.webtrends.corp/guestAuth/repository/download/bt371/.lastSuccessful/action-center-integrations-example-develop-SNAPSHOT-bin.tar.gz"
+default['wt_actioncenter_dd_responsys']['message_port']   = 2552