+++ cookbooks/wt_thumbnailcapture/metadata.rb
@@ -1,0 +1,9 @@
+name                "wt_thumbnailcapture"
+maintainer          "Webtrends, Inc"
+maintainer_email    "toby.mosby@webtrends.com"
+license             "All rights reserved"
+description         "Installs/Configures the Webtrends URL Thumbnail Capture"
+long_description    IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version             "1.0.3"
+depends             "java"
+depends             "runit"