+++ cookbooks/application_ruby/CHANGELOG.md
@@ -1,3 +1,17 @@
+## v1.1.0
+
+* [COOK-2362] - `application_ruby` unicorn uses `run_restart`
+* [COOK-2363] - `application_ruby` unicorn should set
+  `log_template_name` and `run_template_name`
+
+## v1.0.10:
+
+* [COOK-2260] - pin runit version
+
+## v1.0.8:
+
+* [COOK-2159] - cookbook attribute is not treated as a string
+
 ## v1.0.6:
 
 * [COOK-1481] - unicorn provider in application_ruby cookbook should run its restart