+++ cookbooks/wt_portfolio_edge_stream_api/CHANGELOG.md
@@ -1,2 +1,5 @@
+## 1.0.1
+* Bridge changes to support akka
+
 ## 1.0.0
 * Initial release